% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omix.R
\docType{package}
\name{omix}
\alias{omix}
\title{Sample Randomization}
\description{
Chooses the layout that minimizes correlation between
randomization variables and technical covariates.
}
\details{
Provides sample sheets to aid with manual pipetting.
}
\author{
Lucy Sinke \email{l.j.sinke@lumc.nl}
}
