% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omix_corr.R
\name{omix_corr}
\alias{omix_corr}
\title{Correlation Tests}
\usage{
omix_corr(x, y)
}
\arguments{
\item{x}{Randomization variable (e.g. age)}

\item{y}{Technical covariate (e.g. plate number)}
}
\value{
List of correlation estimate and p-value
}
\description{
This function uses appropriate tests of correlation between
two variables and stores the estimate and p-value in a list.
}
\details{
For two categorical variables, the Cramer's V estimate is
stored alongside chi-square p-value. For all other combinations
of variables, Pearson's correlation coefficient and p-value
are stored.

Please note: variables will be converted to numeric class
within this function.
}
\examples{
age <- rnorm(n = 20, mean = 70, sd = 3)
plate <- rep(1:4, each = 5)
omix_corr(age, plate)

}
