% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omix_cv.R
\name{omix_cv}
\alias{omix_cv}
\title{Cramer's V Test}
\usage{
omix_cv(x, y)
}
\arguments{
\item{x}{Randomization variable (e.g. sex)}

\item{y}{Technical covariate (e.g. plate number)}
}
\value{
Cramer's V estimate of correlation
}
\description{
This function calculates the Cramer's V estimate of correlation
between two categorical variables with a value between 0 and 1.
}
\examples{
sex <- sample(1:2, 20, replace = TRUE)
plate <- rep(1:4, each = 5)
omix_corr(sex, plate)

}
