% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omix_rand.R
\name{omix_rand}
\alias{omix_rand}
\title{Sample Randomization}
\usage{
omix_rand(
  df,
  sample_num = sample_num,
  block_var = block_var,
  iter_num = 10000,
  plate_layout = "Illumina96",
  rand_vars,
  tech_vars,
  mask = 0,
  save = "rdata"
)
}
\arguments{
\item{df}{Sample list}

\item{sample_num}{Name of sample ID variable}

\item{block_var}{(Optional) Paired sample identifier}

\item{iter_num}{(Optional) Number of layouts to generate (default: 10,000)}

\item{plate_layout}{Pre-specified or custom plate layout}

\item{rand_vars}{(Optional) Randomization variables (default: all except sample ID and blocking variable)}

\item{tech_vars}{Technical covariates}

\item{mask}{Wells to be left empty}

\item{save}{Method of saving output}
}
\value{
Data frame containing the chosen layout
}
\description{
This function takes a sample list, with defined sample ID,
and randomizes it a specified number of times.
}
\details{
It has an option to specify paired samples (blocks), which
will be kept adjacent to one another but with their order
shuffled.

Randomized sample layouts are then combined with either a
pre-specified or custom plate layout, with the option to
mask specified wells.

The function then chooses the layout that minimizes
correlation between randomization variables and technical
covariates, so long as no test of correlation reached
significance (p<0.05) for that layout.
}
