% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omix_sheet.R
\name{omix_sheet}
\alias{omix_sheet}
\title{Sample Sheet Generation}
\usage{
omix_sheet(layout = final_layout, group)
}
\arguments{
\item{layout}{List of randomized samples}

\item{group}{(Optional) Colour-coding indicator}
}
\value{
PDF of sample layout in working directory
}
\description{
This function takes a sample layout, either specified or
from the omix_rand() function and generates a PDF of the
layout to aid with manual pipetting.
}
