% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omix_specific.R
\name{omix_specific}
\alias{omix_specific}
\title{Sample List Generation}
\usage{
omix_specific(
  df,
  seed = seed,
  sample_id = sample_id,
  block = block,
  plate_layout = "Illumina96",
  rand_vars,
  tech_vars,
  mask = 0,
  save = "rdata"
)
}
\arguments{
\item{df}{Sample list}

\item{seed}{Seed returned by omix_rand()}

\item{sample_id}{Name of sample ID variable}

\item{block}{(Optional) Paired sample identifier}

\item{plate_layout}{Pre-specified or custom plate layout}

\item{rand_vars}{(Optional) Randomization variables (default: all except sample ID and blocking variable)}

\item{tech_vars}{Technical covariates}

\item{mask}{Wells to be left empty}

\item{save}{Method of saving output}
}
\value{
Data frame containing the chosen layout
}
\description{
This function takes a sample list and seed, and returns the
layout that would have been produced by omix_rand() with this seed.
}
\details{
This is for regeneration of randomized sample lists if, for example,
adding new columns that would not affect optimal randomization.
}
