% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixerCorr.R
\name{omixerCorr}
\alias{omixerCorr}
\title{Correlation Tests}
\usage{
omixerCorr(x, y)
}
\arguments{
\item{x}{Randomization variable (e.g. age)}

\item{y}{Technical covariate (e.g. plate number)}
}
\value{
List of correlation estimate and p-value
}
\description{
This function uses appropriate tests of correlation between
two variables and stores the estimate and p-value in a list.
}
\details{
For two categorical variables, the Cramer's V estimate is
stored alongside chi-square p-value. For all other combinations
of variables, Pearson's correlation coefficient and p-value
are stored.

Please note: variables will be converted to numeric class
within this function.
}
