% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixr.R
\docType{package}
\name{omixr}
\alias{omixr}
\title{Multivariate Randomization}
\description{
omixr provides functions to perform optimal randomization of
sample lists prior to omic profiling. This minimizes the
correlation between biological factors and technical covariates,
whilst ensuring there is insufficient evidence for any of
these associations.
}
\details{
Following this multivariate randomization, omixr can also be
used to smoothly bridge the gap between dry and wet labs, by
creating visually intuitive sample sheets.

With clear documentation and the possibility to reproduce any
output sample list, omixr sets the standard for transparency
and reproducibility in this often vague methodological step.
}
\author{
Lucy Sinke \email{l.j.sinke@lumc.nl}
}
