% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixr_rand.R
\name{omixr_rand}
\alias{omixr_rand}
\title{Multivariate Randomization}
\usage{
omixr_rand(
  df,
  sample_id = sample_id,
  block = block,
  iter_num = 10000,
  plate_layout = "Illumina96",
  rand_vars,
  tech_vars,
  mask = 0
)
}
\arguments{
\item{df}{Sample list}

\item{sample_id}{String specifying sample ID variable}

\item{block}{Paired sample identifier}

\item{iter_num}{Number of layouts to generate}

\item{plate_layout}{Pre-specified or custom plate layout}

\item{rand_vars}{Randomization variables}

\item{tech_vars}{Technical covariates}

\item{mask}{Wells to be left empty}
}
\value{
Selected randomized sample list as a data frame
}
\description{
As the main function of the omixr package, this function
outputs a randomized sample list that minimizes correlations
between biological factors and technical covariates.
}
