% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixr_sheet.R
\name{omixr_sheet}
\alias{omixr_sheet}
\title{Sample Sheet Generation}
\usage{
omixr_sheet(omixr_layout = omixr_layout, group = NULL)
}
\arguments{
\item{omixr_layout}{Randomized sample list}

\item{group}{Colour-coding indicator}
}
\value{
PDF of sample layout in working directory
}
\description{
This function will generate visually intuitive plate layouts
for the wet lab, with the option to colour code different types
of samples (e.g. for studies investigating multiple tissues).
}
