% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixr_specific.R
\name{omixr_specific}
\alias{omixr_specific}
\title{Sample List Regeneration}
\usage{
omixr_specific(
  df,
  seed = seed,
  sample_id = sample_id,
  block = block,
  plate_layout = "Illumina96",
  rand_vars,
  tech_vars,
  mask = 0
)
}
\arguments{
\item{df}{Sample list}

\item{seed}{Seed}

\item{sample_id}{Name of sample ID variable}

\item{block}{Paired sample identifier}

\item{plate_layout}{Pre-specified or custom plate layout}

\item{rand_vars}{Randomization variables (default: all except sample ID and block)}

\item{tech_vars}{Technical covariates}

\item{mask}{Wells to be left empty}
}
\value{
Chosen layout as a data frame
}
\description{
Regenerate an omixr-produced randomized sample list quickly,
by providing the seed returned from omixr_rand().
}
