% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixerRand.R
\name{omixerRand}
\alias{omixerRand}
\title{Multivariate Randomization}
\usage{
omixerRand(
  df,
  sampleId = "sampleId",
  block = "block",
  iterNum = 1000,
  wells,
  div = "none",
  positional = FALSE,
  plateNum = 1,
  layout,
  mask = 0,
  techVars,
  randVars
)
}
\arguments{
\item{df}{Sample list}

\item{sampleId}{String specifying sample ID variable}

\item{block}{Paired sample identifier}

\item{iterNum}{Number of layouts to generate}

\item{wells}{Number of wells on a plate}

\item{div}{Plate subdivisions}

\item{positional}{Logical indicator of positional batch effects}

\item{plateNum}{Number of plates}

\item{layout}{Custom plate layout as data frame}

\item{mask}{Wells to be left empty}

\item{techVars}{Technical covariates}

\item{randVars}{Randomization variables}
}
\value{
Selected randomized sample list as a data frame
}
\description{
As the main function of the Omixer package, this function
outputs a randomized sample list that minimizes correlations
between biological factors and technical covariates.
}
