% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixerSheet.R
\name{omixerSheet}
\alias{omixerSheet}
\title{Sample Sheet Generation}
\usage{
omixerSheet(omixerLayout = omixerLayout, group = NULL)
}
\arguments{
\item{omixerLayout}{Randomized sample list}

\item{group}{Colour-coding indicator}
}
\value{
PDF of sample layout in working directory
}
\description{
This function will generate visually intuitive plate layouts
for the wet lab, with the option to colour code different types
of samples (e.g. for studies investigating multiple tissues).
}
