% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixerSpecific.R
\name{omixerSpecific}
\alias{omixerSpecific}
\title{Sample List Regeneration}
\usage{
omixerSpecific(
  df,
  seed = seed,
  sampleId = "sampleId",
  block = "block",
  wells,
  div = "none",
  positional = FALSE,
  plateNum = 1,
  layout,
  mask = 0,
  techVars,
  randVars
)
}
\arguments{
\item{df}{Sample list}

\item{seed}{Seed}

\item{sampleId}{String specifying sample ID variable}

\item{block}{Paired sample identifier}

\item{wells}{Number of wells on a plate}

\item{div}{Plate subdivisions}

\item{positional}{Logical indicator of positional batch effects}

\item{plateNum}{Number of plates}

\item{layout}{Custom plate layout as data frame}

\item{mask}{Wells to be left empty}

\item{techVars}{Technical covariates}

\item{randVars}{Randomization variables}
}
\value{
Chosen layout as a data frame
}
\description{
Regenerate an Omixer-produced randomized sample list quickly,
by providing the seed returned from omixerRand().
}
